% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Samples.R
\name{makeFakeTable}
\alias{makeFakeTable}
\title{makeFakeTable}
\usage{
makeFakeTable()
}
\value{
A fake table of type data.table to play around with
}
\description{
Makes a table to play around with
}
\examples{
makeFakeTable()

}
