% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreProcessing.R
\name{makeNiceList}
\alias{makeNiceList}
\title{makeNiceList}
\usage{
makeNiceList(grossList)
}
\arguments{
\item{grossList}{the grossly formated list}
}
\value{
a list of symbols that aren't themselves also lists.
}
\description{
Takes a gross variable length list and converts it into a nice one for use
in proportionalize
}
\examples{
makeNiceList(theColumnsVariableFromProportionalize)

}
