% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreProcessing.R
\name{proportionalize}
\alias{proportionalize}
\title{proportionalize}
\usage{
proportionalize(theTable, priCol, ...)
}
\arguments{
\item{theTable}{The table whose columns are being manipulated}

\item{priCol}{The column which will be the one the output column
will be the proportion of}

\item{cols}{The columns being aggregated}
}
\value{
The table with its columns aggregated
}
\description{
Takes the provided columns and collapses them into one column which is 
the proportion of the time spent in the selected category
}
\examples{
proportionalize(funTable, funtime, seriousTime, funTime)

}
