% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreProcessing.R
\name{stripUseless}
\alias{stripUseless}
\title{stripUseless}
\usage{
stripUseless(theTable, ...)
}
\arguments{
\item{...}{A list of columns to remove}

\item{table}{The table being stripped}
}
\value{
a modified table which no longer has the undesired columns
}
\description{
Strips columns that are most likely useless such as
"Browser" and "OperatingSystems"
}
\examples{
stripUseless (table)

stripUseless(data.table(x = c(1,2,3), y = c(1,2,3)), x)


}
